Randomly driving maze

+++ ws/src/micromouse_robot/include/micromouse_robot/pid_controller.hpp
@@ -3,16 +3,16 @@
 namespace micromouse{
 class PidController {
    public:
-    PidController(double kp, double kd, double ki);
-    double get_command(double command, double state, double time);
+    PidController(float kp, float kd, float ki);
+    float get_command(float command, float state, float time);
 
    private:
-    double kp_;
-    double kd_;
-    double ki_;
+    float kp_;
+    float kd_;
+    float ki_;
 
-    double last_error_;
-    double integral_;
-    double last_time_;
+    float last_error_;
+    float integral_;
+    float last_time_;
 };
 }  // namespace
